refactor: replace Resource type with enum
