+++ src/app/guide_articles/layout.tsx
@@ -4,15 +4,15 @@ import { usePathname } from "next/navigation";
 import type { ReactNode } from "react";
 
 import { AbstractResourceLayout } from "@/components/abstract/abstract-resource-layout";
-import { ResourcePaths } from "@/lib/enums";
+import { Resource } from "@/lib/enums";
 
 export default function Layout({ children }: { children: ReactNode }) {
   const pathname = usePathname();
 
   const titleMap = {
-    [`/${ResourcePaths.GuideArticles}/create`]: "Dodawanie artykułu",
-    [`/${ResourcePaths.GuideArticles}/edit`]: "Edycja artykułu",
-    [`/${ResourcePaths.GuideArticles}`]: "Zarządzenie artykułami",
+    [`/${Resource.GuideArticles}/create`]: "Dodawanie artykułu",
+    [`/${Resource.GuideArticles}/edit`]: "Edycja artykułu",
+    [`/${Resource.GuideArticles}`]: "Zarządzenie artykułami",
   };
 
   return (
