+++ src/app/guide_articles/page.tsx
@@ -1,5 +1,5 @@
 import { AbstractResourceList } from "@/components/abstract/abstract-resource-list";
-import { ResourcePaths } from "@/lib/enums";
+import { Resource } from "@/lib/enums";
 import type { GuideArticle } from "@/types/app";
 
 export default function GuideArticlesPage({
@@ -9,7 +9,7 @@ export default function GuideArticlesPage({
 }) {
   return (
     <AbstractResourceList
-      resource={ResourcePaths.GuideArticles}
+      resource={Resource.GuideArticles}
       searchParams={searchParams}
       mapItemToList={(item: GuideArticle) => ({
         id: item.id,
