+++ src/app/student_organizations/edit/[id]/page.tsx
@@ -1,5 +1,5 @@
 import { AbstractResourceEditPage } from "@/components/abstract/abstract-resource-edit-page";
-import { ResourcePaths } from "@/lib/enums";
+import { Resource } from "@/lib/enums";
 
 import { Form } from "../../form";
 
@@ -10,7 +10,7 @@ export default function EditStudentOrganizationPage({
 }) {
   return (
     <AbstractResourceEditPage
-      resource={ResourcePaths.StudentOrganizations}
+      resource={Resource.StudentOrganizations}
       params={params}
       FormComponent={Form}
     />
