+++ src/app/student_organizations/layout.tsx
@@ -4,15 +4,15 @@ import { usePathname } from "next/navigation";
 import type { ReactNode } from "react";
 
 import { AbstractResourceLayout } from "@/components/abstract/abstract-resource-layout";
-import { ResourcePaths } from "@/lib/enums";
+import { Resource } from "@/lib/enums";
 
 export default function Layout({ children }: { children: ReactNode }) {
   const pathname = usePathname();
 
   const titleMap = {
-    [`/${ResourcePaths.StudentOrganizations}/create`]: "Dodawanie organizacji",
-    [`/${ResourcePaths.StudentOrganizations}/edit`]: "Edycja organizacji",
-    [`/${ResourcePaths.StudentOrganizations}`]: "Zarządzanie organizacjami",
+    [`/${Resource.StudentOrganizations}/create`]: "Dodawanie organizacji",
+    [`/${Resource.StudentOrganizations}/edit`]: "Edycja organizacji",
+    [`/${Resource.StudentOrganizations}`]: "Zarządzanie organizacjami",
   };
 
   return (
