+++ src/lib/enums.ts
@@ -1,4 +1,4 @@
-export enum ResourcePaths {
+export enum Resource {
   GuideArticles = "guide_articles",
   StudentOrganizations = "student_organizations",
 }
